Change cli default printing order
When printing all the data for the circuit, the data will be printed
always in the same order (currents, component voltage, node voltage,
continuous states, forward, backward and trapezoidal). When adding
multiple flags, there is also a default order, but it was different
(the first three where nodes, component voltages and currents). Now, the
order for printing with flags is the same as the "print all" one.